feat: finish with preview
